+++ __tests__/Routes.test.tsx
@@ -1,6 +1,8 @@
 import { defineComponent, Fragment, nextTick, onMounted } from 'vue'
 import { mount } from '@vue/test-utils'
-import { MemoryRouter as Router, Routes, Route, useNavigate } from '@'
+import { MemoryRouter as Router, Route, useNavigate } from '@'
+
+import Routes from './utils/RoutesDisableKeepAlive'
 
 describe('A <Routes>', () => {
     it('renders the first route that matches the URL', () => {