+++ __tests__/utils/RoutesDisableKeepAlive.tsx
@@ -1,0 +1,5 @@
+import { Routes } from '@'
+
+export default (props: any, { slots }: any) => {
+    return <Routes {...props} roughlyDisableKeepAlive v-slots={slots} />
+}